requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
async==0.6.2
beautifulsoup4==4.4.1
billiard==3.5.0.4
celery==4.2.0
certifi==2018.1.18
cffi==1.11.5
cryptography==2.3
Django==1.11.14
django-bootstrap3==9.1.0
django-celery-results==1.0.1
django-haystack==2.7.0
django-redis==4.7.0
django-tinymce==2.7.0
djangorestframework==3.8.2
eventlet==0.23.0
Flask==0.10.1
Flask-SQLAlchemy==2.0
greenlet==0.4.14
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jieba==0.39
Jinja2==2.7.3
kombu==4.2.1
MarkupSafe==0.23
Naked==0.1.31
Pillow==5.2.0
pycparser==2.18
pycryptodomex==3.6.4
PyMySQL==0.9.2
python-alipay-sdk==1.7.1
pytz==2015.6
PyYAML==3.13
redis==2.10.6
requests==2.7.0
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.0.8
vine==1.1.4
Werkzeug==0.10.4
Whoosh==2.7.4
